+++ simplesockets/__init__.py
@@ -1,9 +1,3 @@
 from .simple_sockets import TCPServer, TCPClient
-from .securesockets.SelfmadeCipher_Socket import SecureClient as PySecureClient, SecureServer as PySecureServer
-from .securesockets.PyCryptodome_Socket import SecureClient, SecureServer
-
-try:
-    from Crypto.PublicKey.RSA import RsaKey
-    PYCRYPTODOME = True
-except ImportError:
-    PYRCYPTODOME = False
+from .secure_sockets import SecureServer, SecureClient
+from . import typehints