+++ simplesockets/typehints.py
@@ -1,0 +1,3 @@
+from .simple_sockets import Socket_Response, Server_Client
+from ._support_files.Events import Event
+from ._support_files.error import error, SocketError, SetupError, Better_Exception